feat(typeAhead): Add soft autocomplete feature and allow user to configure it using typeAhead prop
BREAKING CHANGE: Finally added soft autocomplete feature. Major version update!
