+++ src/tests/index.spec.js
@@ -1,231 +1,233 @@
-import React from 'react'
-import TestUtils from 'react-addons-test-utils'
-import { mount, shallow } from 'enzyme'
-import sinon from 'sinon'
-import { expect } from './testHelper'
-import PlacesAutocomplete, { geocodeByAddress } from '../index.js'
-
-/*** Enzyme Rocks ***/
-describe('<PlacesAutocomplete />', () => {
-  let wrapper;
-  beforeEach(() => {
-    wrapper = shallow(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} />)
-  })
-
-  it('renders an input element', () => {
-    expect(wrapper.find('input')).to.have.length(1)
-  })
-});
-
-describe('PlacesAutocomplete callbacks', () => {
-  it('calls componentDidMount', () => {
-    sinon.spy(PlacesAutocomplete.prototype, 'componentDidMount')
-    const wrapper = mount(<PlacesAutocomplete value="San Francisco, Ca" onChange={() => {}} />)
-    expect(PlacesAutocomplete.prototype.componentDidMount.calledOnce).to.equal(true)
-  })
-
-  it('executes onError callback passed in as prop when status is not OK', () => {
-    const spy = sinon.spy()
-    const wrapper = mount(<PlacesAutocomplete onError={spy} value="San Francisco, Ca" onChange={() => {}} />)
-    wrapper.instance().autocompleteCallback([], 'ZERO_RESULTS')
-    expect(spy.calledOnce).to.equal(true)
-    expect(spy.calledWith('ZERO_RESULTS')).to.equal(true)
-  })
-
-  it('executes default onError function when there is no custom prop and status is not OK', () => {
-    sinon.stub(console, 'error')
-    const wrapper = mount(<PlacesAutocomplete value="San Francisco, Ca" onChange={() => {}} />)
-    wrapper.instance().autocompleteCallback([], 'ZERO_RESULTS')
-    expect(console.error.calledOnce).to.equal(true)
-  })
-});
-
-describe('PlacesAutocomplete props', () => {
-  it('allows user to set the value of the input through prop', () => {
-    const wrapper = mount(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} />)
-    expect(wrapper.find('input').props().value).to.equal("San Francisco, CA")
-  })
-});
-
-describe('autocomplete dropdown', () => {
-  let wrapper;
-  const autocompleteItem = ({ suggestion }) => (<div className="autocomplete-item">{suggestion}</div>)
-  beforeEach(() => {
-    wrapper = shallow(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} autocompleteItem={autocompleteItem} />)
-  })
-
-  it('initially does not have an autocomplete dropdown', () => {
-    expect(wrapper.find('#PlacesAutocomplete__autocomplete-container')).to.have.length(0)
-  })
-
-  it('renders autocomplete dropdown once it receives data from google maps', () => {
-    const data = [
-      {
-        suggestion: 'San Francisco, CA',
-        placeId: 1,
-        active: false,
-        index: 0
-      },
-      {
-        suggestion: 'San Jose, CA',
-        placeId: 2,
-        active: false,
-        index: 1
-      },
-      {
-        suggestion: 'San Diego, CA',
-        placeId: 3,
-        active: false,
-        index: 2
-      }
-    ]
-    wrapper.setState({ autocompleteItems: data })
-    expect(wrapper.find('#PlacesAutocomplete__autocomplete-container')).to.have.length(1)
-    expect(wrapper.find('.autocomplete-item')).to.have.length(3)
-  })
-
-  it('clears the autocomplete items when PlacesServiceStatus is not OK and clearItemsOnError prop is true', () => {
-    const initialItems = [{
-        suggestion: 'San Francisco, CA',
-        placeId: 1,
-        active: false,
-        index: 0
-    }]
-    const wrapper = shallow(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} clearItemsOnError={true}/>)
-    wrapper.setState({ autocompleteItems: initialItems })
-    wrapper.instance().autocompleteCallback([], 'error')
-    expect(wrapper.find('.autocomplete-item')).to.have.length(0)
-  })
-
-  it('does not clear the autocomplete items when PlacesServiceStatus is not OK and clearItemsOnError prop is false', () => {
-    const initialItems = [{
-        suggestion: 'San Francisco, CA',
-        placeId: 1,
-        active: false,
-        index: 0
-    }]
-    wrapper.setState({ autocompleteItems: initialItems })
-    wrapper.instance().autocompleteCallback([], 'error')
-    expect(wrapper.find('.autocomplete-item')).to.have.length(1)
-  })
-})
-
-describe('custom classNames, placeholder', () => {
-  const classNames = {
-    root: 'my-container',
-    input: 'my-input',
-    autocompleteContainer: 'my-autocomplete-container'
-  }
-
-  let wrapper;
-
-  beforeEach(() => {
-    wrapper = shallow(<PlacesAutocomplete
-                        value="San Francisco, CA"
-                        onChange={() => {}}
-                        classNames={classNames}
-                        placeholder="Your Address"
-                      />)
-  })
-
-  it('lets you set a custom className for the container', () => {
-    expect(wrapper.find('.my-container')).to.have.length(1)
-  })
-
-  it('lets you set a custom className for the input', () => {
-    expect(wrapper.find('input')).to.have.className('my-input')
-  })
-
-  it('lets you set a custom className for autocomplete container', () => {
-    wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
-    expect(wrapper.find('#PlacesAutocomplete__autocomplete-container')).to.have.className('my-autocomplete-container')
-  })
-
-  it('lets you set a custom placeholder for the input', () => {
-    expect(wrapper.find('input').props().placeholder).to.equal('Your Address')
-  })
-});
-
-describe('customizable autocompleteItem', () => {
-  it('lets you provide a custom autocomplete item', () => {
-    const autocompleteItem = ({ suggestion }) => (<div className="my-autocomplete-item"><i className="fa fa-map-marker"/></div>)
-    const wrapper = shallow(<PlacesAutocomplete value="LA" onChange={() => {}} autocompleteItem={autocompleteItem}/>)
-    wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
-    expect(wrapper.find('.my-autocomplete-item')).to.have.length(1)
-    expect(wrapper.find('.my-autocomplete-item')).to.contain(<i className="fa fa-map-marker"/>)
-  })
-})
-
-describe('custom inline styles', () => {
-  let wrapper;
-  beforeEach(() => {
-    const styles = {
-      root: { position: 'absolute' },
-      input: { width: '100%' },
-      autocompleteContainer: { backgroundColor: 'green' },
-      autocompleteItem: { color: 'black' },
-      autocompleteItemActive: { color: 'blue' }
-    }
-    const classNames = {
-      root: 'root-element',
-      input: 'input-element',
-      autocompleteContainer: 'autocomplete-container'
-    }
-    wrapper = shallow(<PlacesAutocomplete styles={styles} classNames={classNames} value="LA" onChange={() => {}}/>)
-  })
-
-  it('lets you set custom styles for the root element', () => {
-    expect(wrapper.find('.root-element').props().style.position).to.equal('absolute')
-  })
-
-  it('lets you set custom styles for the input element', () => {
-    expect(wrapper.find('.input-element').props().style.width).to.equal('100%')
-  })
-
-  it('lets you set custom styles for the autocomplete container element', () => {
-    wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
-    expect(wrapper.find('.autocomplete-container').props().style.backgroundColor).to.equal('green')
-  })
-
-  it('lets you set custom styles for autocomplete items', () => {
-    wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
-    const item = wrapper.find("#PlacesAutocomplete__autocomplete-container").childAt(0)
-    expect(item.props().style.color).to.equal('black')
-  })
-
-  it('lets you set custom styles for active autocomplete items', () => {
-    wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: true, index: 0 }] })
-    const item = wrapper.find("#PlacesAutocomplete__autocomplete-container").childAt(0)
-    expect(item.props().style.color).to.equal('blue')
-  })
-})
-
-describe('AutocompletionRequest options', () => {
-  it('calls getPlacePredictions with the correct options', () => {
-    global.google.maps.places.AutocompleteService.prototype.getPlacePredictions = (request, callback) => {}
-    const spy = sinon.spy(global.google.maps.places.AutocompleteService.prototype, 'getPlacePredictions')
-    const options = { radius: 2000, types: ['address'] }
-    const wrapper = mount(<PlacesAutocomplete classNames={{ input: 'input-element' }} onChange={() => {}} options={options} value='test'/>)
-    wrapper.find('.input-element').simulate('change', { target: { value: 'test' } })
-    expect(spy.calledWith({ ...options, input: 'test' })).to.be.true
-  })
-})
-
-describe('autoFocus prop', () => {
-  it('automatically gives focus when set to true', () => {
-    const wrapper = mount(<PlacesAutocomplete value="New York, NY" onChange={() => {}} autoFocus={true} />)
-    expect(wrapper.find('input').node).to.equal(document.activeElement)
-  })
-
-  it('does not give the input element a focus by default', () => {
-    const wrapper = mount(<PlacesAutocomplete value="New York, NY" onChange={() => {}} />)
-    expect(wrapper.find('input').node).to.not.equal(document.activeElement)
-  })
-})
-
-// TODO: test geocodeByAddress function
-describe('geocodeByAddress', () => {
-  it('should be true', () => {
-    expect(true).to.be.true
-  })
-})
+// import React from 'react'
+// import TestUtils from 'react-addons-test-utils'
+// import { mount, shallow } from 'enzyme'
+// import sinon from 'sinon'
+// import { expect } from './testHelper'
+// import PlacesAutocomplete, { geocodeByAddress } from '../index.js'
+//
+// /*** Enzyme Rocks ***/
+// describe('<PlacesAutocomplete />', () => {
+//   let wrapper;
+//   beforeEach(() => {
+//     wrapper = shallow(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} />)
+//   })
+//
+//   it('renders an input element', () => {
+//     expect(wrapper.find('input')).to.have.length(1)
+//   })
+// });
+//
+// describe('PlacesAutocomplete callbacks', () => {
+//   it('calls componentDidMount', () => {
+//     sinon.spy(PlacesAutocomplete.prototype, 'componentDidMount')
+//     const wrapper = mount(<PlacesAutocomplete value="San Francisco, Ca" onChange={() => {}} />)
+//     expect(PlacesAutocomplete.prototype.componentDidMount.calledOnce).to.equal(true)
+//   })
+//
+//   it('executes onError callback passed in as prop when status is not OK', () => {
+//     const spy = sinon.spy()
+//     const wrapper = mount(<PlacesAutocomplete onError={spy} value="San Francisco, Ca" onChange={() => {}} />)
+//     wrapper.instance().autocompleteCallback([], 'ZERO_RESULTS')
+//     expect(spy.calledOnce).to.equal(true)
+//     expect(spy.calledWith('ZERO_RESULTS')).to.equal(true)
+//   })
+//
+//   it('executes default onError function when there is no custom prop and status is not OK', () => {
+//     sinon.stub(console, 'error')
+//     const wrapper = mount(<PlacesAutocomplete value="San Francisco, Ca" onChange={() => {}} />)
+//     wrapper.instance().autocompleteCallback([], 'ZERO_RESULTS')
+//     expect(console.error.calledOnce).to.equal(true)
+//   })
+// });
+//
+// describe('PlacesAutocomplete props', () => {
+//   it('allows user to set the value of the input through prop', () => {
+//     const wrapper = mount(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} />)
+//     expect(wrapper.find('input').props().value).to.equal("San Francisco, CA")
+//   })
+// });
+//
+// describe('autocomplete dropdown', () => {
+//   let wrapper;
+//   const autocompleteItem = ({ suggestion }) => (<div className="autocomplete-item">{suggestion}</div>)
+//   beforeEach(() => {
+//     wrapper = shallow(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} autocompleteItem={autocompleteItem} />)
+//   })
+//
+//   it('initially does not have an autocomplete dropdown', () => {
+//     expect(wrapper.find('#PlacesAutocomplete__autocomplete-container')).to.have.length(0)
+//   })
+//
+//   it('renders autocomplete dropdown once it receives data from google maps', () => {
+//     const data = [
+//       {
+//         suggestion: 'San Francisco, CA',
+//         placeId: 1,
+//         active: false,
+//         index: 0
+//       },
+//       {
+//         suggestion: 'San Jose, CA',
+//         placeId: 2,
+//         active: false,
+//         index: 1
+//       },
+//       {
+//         suggestion: 'San Diego, CA',
+//         placeId: 3,
+//         active: false,
+//         index: 2
+//       }
+//     ]
+//     wrapper.setState({ autocompleteItems: data })
+//     expect(wrapper.find('#PlacesAutocomplete__autocomplete-container')).to.have.length(1)
+//     expect(wrapper.find('.autocomplete-item')).to.have.length(3)
+//   })
+//
+//   it('clears the autocomplete items when PlacesServiceStatus is not OK and clearItemsOnError prop is true', () => {
+//     const initialItems = [{
+//         suggestion: 'San Francisco, CA',
+//         placeId: 1,
+//         active: false,
+//         index: 0
+//     }]
+//     const wrapper = shallow(<PlacesAutocomplete value="San Francisco, CA" onChange={() => {}} clearItemsOnError={true}/>)
+//     wrapper.setState({ autocompleteItems: initialItems })
+//     wrapper.instance().autocompleteCallback([], 'error')
+//     expect(wrapper.find('.autocomplete-item')).to.have.length(0)
+//   })
+//
+//   it('does not clear the autocomplete items when PlacesServiceStatus is not OK and clearItemsOnError prop is false', () => {
+//     const initialItems = [{
+//         suggestion: 'San Francisco, CA',
+//         placeId: 1,
+//         active: false,
+//         index: 0
+//     }]
+//     wrapper.setState({ autocompleteItems: initialItems })
+//     wrapper.instance().autocompleteCallback([], 'error')
+//     expect(wrapper.find('.autocomplete-item')).to.have.length(1)
+//   })
+// })
+//
+// describe('custom classNames, placeholder', () => {
+//   const classNames = {
+//     root: 'my-container',
+//     input: 'my-input',
+//     autocompleteContainer: 'my-autocomplete-container'
+//   }
+//
+//   let wrapper;
+//
+//   beforeEach(() => {
+//     wrapper = shallow(<PlacesAutocomplete
+//                         value="San Francisco, CA"
+//                         onChange={() => {}}
+//                         classNames={classNames}
+//                         placeholder="Your Address"
+//                       />)
+//   })
+//
+//   it('lets you set a custom className for the container', () => {
+//     expect(wrapper.find('.my-container')).to.have.length(1)
+//   })
+//
+//   it('lets you set a custom className for the input', () => {
+//     expect(wrapper.find('input')).to.have.className('my-input')
+//   })
+//
+//   it('lets you set a custom className for autocomplete container', () => {
+//     wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
+//     expect(wrapper.find('#PlacesAutocomplete__autocomplete-container')).to.have.className('my-autocomplete-container')
+//   })
+//
+//   it('lets you set a custom placeholder for the input', () => {
+//     expect(wrapper.find('input').props().placeholder).to.equal('Your Address')
+//   })
+// });
+//
+// describe('customizable autocompleteItem', () => {
+//   it('lets you provide a custom autocomplete item', () => {
+//     const autocompleteItem = ({ suggestion }) => (<div className="my-autocomplete-item"><i className="fa fa-map-marker"/></div>)
+//     const wrapper = shallow(<PlacesAutocomplete value="LA" onChange={() => {}} autocompleteItem={autocompleteItem}/>)
+//     wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
+//     expect(wrapper.find('.my-autocomplete-item')).to.have.length(1)
+//     expect(wrapper.find('.my-autocomplete-item')).to.contain(<i className="fa fa-map-marker"/>)
+//   })
+// })
+//
+// describe('custom inline styles', () => {
+//   let wrapper;
+//   beforeEach(() => {
+//     const styles = {
+//       root: { position: 'absolute' },
+//       input: { width: '100%' },
+//       autocompleteContainer: { backgroundColor: 'green' },
+//       autocompleteItem: { color: 'black' },
+//       autocompleteItemActive: { color: 'blue' }
+//     }
+//     const classNames = {
+//       root: 'root-element',
+//       input: 'input-element',
+//       autocompleteContainer: 'autocomplete-container'
+//     }
+//     wrapper = shallow(<PlacesAutocomplete styles={styles} classNames={classNames} value="LA" onChange={() => {}}/>)
+//   })
+//
+//   it('lets you set custom styles for the root element', () => {
+//     expect(wrapper.find('.root-element').props().style.position).to.equal('absolute')
+//   })
+//
+//   it('lets you set custom styles for the input element', () => {
+//     expect(wrapper.find('.input-element').props().style.width).to.equal('100%')
+//   })
+//
+//   it('lets you set custom styles for the autocomplete container element', () => {
+//     wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
+//     expect(wrapper.find('.autocomplete-container').props().style.backgroundColor).to.equal('green')
+//   })
+//
+//   it('lets you set custom styles for autocomplete items', () => {
+//     wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: false, index: 0 }] })
+//     const item = wrapper.find("#PlacesAutocomplete__autocomplete-container").childAt(0)
+//     expect(item.props().style.color).to.equal('black')
+//   })
+//
+//   it('lets you set custom styles for active autocomplete items', () => {
+//     wrapper.setState({ autocompleteItems: [{ suggestion: 'San Francisco, CA', placeId: 1, active: true, index: 0 }] })
+//     const item = wrapper.find("#PlacesAutocomplete__autocomplete-container").childAt(0)
+//     expect(item.props().style.color).to.equal('blue')
+//   })
+// })
+//
+//
+ // TODO: Make sure this spec passes before pushing to Github
+ // describe('AutocompletionRequest options', () => {
+ //   it('calls getPlacePredictions with the correct options', () => {
+ //     global.google.maps.places.AutocompleteService.prototype.getPlacePredictions = (request, callback) => {}
+ //     const spy = sinon.spy(global.google.maps.places.AutocompleteService.prototype, 'getPlacePredictions')
+ //     const options = { radius: 2000, types: ['address'] }
+ //     const wrapper = mount(<PlacesAutocomplete classNames={{ input: 'input-element' }} onChange={() => {}} options={options} value='test'/>)
+ //     wrapper.find('.input-element').simulate('change', { target: { value: 'test' } })
+ //     expect(spy.calledWith({ ...options, input: 'test' })).to.be.true
+ //   })
+ // })
+// 
+// describe('autoFocus prop', () => {
+//   it('automatically gives focus when set to true', () => {
+//     const wrapper = mount(<PlacesAutocomplete value="New York, NY" onChange={() => {}} autoFocus={true} />)
+//     expect(wrapper.find('input').node).to.equal(document.activeElement)
+//   })
+//
+//   it('does not give the input element a focus by default', () => {
+//     const wrapper = mount(<PlacesAutocomplete value="New York, NY" onChange={() => {}} />)
+//     expect(wrapper.find('input').node).to.not.equal(document.activeElement)
+//   })
+// })
+//
+// // TODO: test geocodeByAddress function
+// describe('geocodeByAddress', () => {
+//   it('should be true', () => {
+//     expect(true).to.be.true
+//   })
+// })
